// src/models/Activity.ts
export interface Activity {
    id: string
    deviceId: string
    monitorId: string
    deviceName: string
    activityContent: any
    activityName: any
    activityContentType: string
    status: string
    command: string
    commandResult: string
    date: Date
    error: string
    type: string
}